requirements.txt
# Core dependencies for Hybrid AI Brain

# Numerical operations and data structures
numpy>=1.24.0
scipy>=1.10.0
networkx>=3.0
pandas>=2.0.0

# Graph Neural Network implementation
torch>=2.0.0
torch-geometric>=2.6.0

# Agentic frameworks integration (updated)

# Use official Microsoft AutoGen
autogen-agentchat>=0.4,<0.5

# Or, if sticking with pyautogen (community fork)
# pyautogen>=0.2.35

langgraph>=0.4.8
langchain>=0.3.0


# Memory and retrieval components
chromadb>=0.4.0
sentence-transformers>=4.1.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0
tqdm>=4.65.0

# Visualization tools
matplotlib>=3.6.0
seaborn>=0.12.0
